refactor(wash-runtime): move config/secret source resolution out of the CLI
`ConfigSource`/`SecretSource` and their `dir:`/`file:`/`inline:`/`fromEnv:`
resolution — path containment checks, kubelet symlink-farm handling, secret
file permission checks, in-repo-tree warnings — lived in `wash`'s CLI crate,
`pub(crate)` (`resolve_environment_layer`) or fully private (`resolve_dir`)
at that. Any other embedder resolving the same `configFrom`/`secretFrom`
references — a Kubernetes control plane, a custom supervisor — had no way to
reuse this and would have to hand-copy it instead: the same class of drift
`with_multiplexed_plugins()` was introduced to kill for plugin registration.

Move the whole source model into `wash_runtime::config_source`, so every
embedder that already depends on the runtime gets it for free.
`wash::config` re-exports `ConfigSource`/`SecretSource`/`EnvironmentLayer`
unchanged so existing callers are unaffected; `HostPluginConfig::to_spec`
now resolves through the runtime's copy instead of its own.

Also drops `wash`'s now-dead `libc` dependency (the moved code uses
`rustix`, already a wash-runtime dependency) and fixes an alphabetical
ordering slip in wash-runtime's Cargo.toml.
